SSH Tunnel:
- Move our SSH tunnel DAO, commands, model and exceptions to its own folder inside databases directory
- Adjust some imports with the new paths
- Add tests to our Delete command
